Plot features for spatial data

The features can include expression, dimension reduction components, metadata, etc

The features can include expression, dimension reduction components, metadata, etc

Usage

SpatialPlot(object, ...)

# S3 method for class 'Seurat'
SpatialPlot(object, image = NULL, ...)

# S3 method for class 'Seurat.VisiumV2'
SpatialPlot(
  object,
  image = NULL,
  masks = NULL,
  shapes = NULL,
  points = NULL,
  ext = NULL,
  image_scale = NULL,
  crop = TRUE,
  group_by = NULL,
  features = NULL,
  layer = "data",
  layers = NULL,
  flip_y = TRUE,
  theme = "theme_box",
  theme_args = list(),
  label = FALSE,
  label_size = 4,
  label_fg = "white",
  label_bg = "black",
  label_bg_r = 0.1,
  label_repel = FALSE,
  label_repulsion = 20,
  label_pt_size = 1,
  label_pt_color = "black",
  label_segment_color = "black",
  label_insitu = FALSE,
  palette = NULL,
  palette_reverse = FALSE,
  palcolor = NULL,
  highlight = NULL,
  highlight_alpha = 1,
  highlight_size = 1,
  highlight_color = "black",
  highlight_stroke = 0.8,
  legend.position = "right",
  legend.direction = "vertical",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  facet_scales = "fixed",
  facet_nrow = NULL,
  facet_ncol = NULL,
  facet_byrow = TRUE,
  ...
)

# S3 method for class 'Seurat.SlideSeq'
SpatialPlot(
  object,
  image = NULL,
  masks = NULL,
  shapes = NULL,
  points = NULL,
  ext = NULL,
  image_scale = NULL,
  crop = TRUE,
  group_by = NULL,
  features = NULL,
  layer = "data",
  layers = NULL,
  flip_y = TRUE,
  theme = "theme_box",
  theme_args = list(),
  label = FALSE,
  label_size = 4,
  label_fg = "white",
  label_bg = "black",
  label_bg_r = 0.1,
  label_repel = FALSE,
  label_repulsion = 20,
  label_pt_size = 1,
  label_pt_color = "black",
  label_segment_color = "black",
  label_insitu = FALSE,
  palette = NULL,
  palette_reverse = FALSE,
  palcolor = NULL,
  highlight = NULL,
  highlight_alpha = 1,
  highlight_size = 1,
  highlight_color = "black",
  highlight_stroke = 0.8,
  legend.position = "right",
  legend.direction = "vertical",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  facet_scales = "fixed",
  facet_nrow = NULL,
  facet_ncol = NULL,
  facet_byrow = TRUE,
  ...
)

SpatialFeaturePlot(object, ...)

# S3 method for class 'Seurat'
SpatialFeaturePlot(object, image = NULL, ...)

SpatialDimPlot(object, ...)

# S3 method for class 'Seurat'
SpatialDimPlot(object, image = NULL, group_by = NULL, ...)

Arguments

object

A Seurat object or a Giotto object.

...

Additional arguments passed to the plotting functions.

image

The name of the image to plot. If NULL, the first image in the object will be used.

Value

A ggplot object

A ggplot object

A ggplot object