Helper functions to select clones based on various criteria

These helper functions allow for the selection of clones based on various criteria such as size, group comparison, and existence in specific groups.

Usage

.get_data(data)

.return_what(data, id, return_all)

top(
  n,
  groups = NULL,
  data = NULL,
  order = NULL,
  id = "CTaa",
  return_all = NULL
)

sel(expr, groups = NULL, data = NULL, id = "CTaa", return_all = NULL)

uniq(
  group1,
  group2,
  ...,
  groups = NULL,
  data = NULL,
  id = "CTaa",
  return_all = NULL
)

shared(
  group1,
  group2,
  ...,
  groups = NULL,
  data = NULL,
  id = "CTaa",
  return_all = NULL
)

gt(
  group1,
  group2,
  include_zeros = TRUE,
  groups = NULL,
  data = NULL,
  id = "CTaa",
  return_all = NULL
)

ge(
  group1,
  group2,
  include_zeros = TRUE,
  groups = NULL,
  data = NULL,
  id = "CTaa",
  return_all = NULL
)

lt(
  group1,
  group2,
  include_zeros = TRUE,
  groups = NULL,
  data = NULL,
  id = "CTaa",
  return_all = NULL
)

le(
  group1,
  group2,
  include_zeros = TRUE,
  groups = NULL,
  data = NULL,
  id = "CTaa",
  return_all = NULL
)

eq(group1, group2, groups = NULL, data = NULL, id = "CTaa", return_all = NULL)

ne(
  group1,
  group2,
  include_zeros = TRUE,
  groups = NULL,
  data = NULL,
  id = "CTaa",
  return_all = NULL
)

and(x, y)

or(x, y)

Arguments

data

The data frame containing clone information. Default is NULL. If NULL, it will get data from parent.frame. A typical data should have a column named CTaa and other columns for the groupings. Supposingly it should be a grouped data frame with the grouping columns. Under each grouping column, the value should be the size of the clone. By default, the data is assumed to be in the parent frame.

id

The column name that contains the clone ID. Default is "CTaa".

return_all

If TRUE, the function returns a vector with the same length as the data, with CTaa values for selected clones and NA for others. If FALSE, it returns a subset data frame with only the selected clones. Default is NULL, which will be determined based on the data. If the function is used in a context of dplyr verbs, it defaults to TRUE. Otherwise, it defaults to FALSE.

n

The number of top clones to select or the threshold size.

groups

The column names in the meta data to group the cells. By default, it is assumed facet_by and split_by to be in the parent frame.

expr

The expression (in characters) to filter the clones (e.g. "group1 > group2" to select clones where group1 is larger than group2).

group1

The first group to compare.

group2

The second group to compare.

...

More groups to compare.

include_zeros

Whether to include clones with zero size in the comparison. If TRUE, in a comparison (s1 > s2) for a clone to be selected, both s1 and s2 must be greater than 0. If FALSE, only the first group must be greater than the second group.

x

The first vector to compare in logical operations (and/or).

y

The second vector to compare in logical operations (and/or).

Value

A vector of CTaas or a data frame with the selected clones based on the criteria.

Examples

data <- data.frame(
    CTaa = 1:10,
    group1 = c(0, 1, 2, 3, 4, 5, 6, 7, 8, 9),
    group2 = c(7, 3, 8, 2, 1, 5, 9, 4, 6, 0),
    groups = c("A", "A", "A", "A", "B", "B", "B", "B", "B", "B")
)
data <- data[order(data$group1 + data$group2, decreasing = TRUE), ]
scplotter:::top(3)
#>   CTaa group1 group2 groups
#> 1    7      6      9      B
#> 2    9      8      6      B
#> 3    8      7      4      B
scplotter:::top(3, groups = "groups")
#> # A tibble: 6 × 4
#> # Groups:   groups [2]
#>    CTaa group1 group2 groups
#>   <int>  <dbl>  <dbl> <chr> 
#> 1     7      6      9 B     
#> 2     9      8      6 B     
#> 3     8      7      4 B     
#> 4     3      2      8 A     
#> 5     1      0      7 A     
#> 6     4      3      2 A     
scplotter:::sel(group1 == 0 | group2 == 0)
#>   CTaa group1 group2 groups
#> 1   10      9      0      B
#> 2    1      0      7      A
scplotter:::uniq(group1, group2)
#>   CTaa group1 group2 groups
#> 1   10      9      0      B
scplotter:::shared(group1, group2)
#>   CTaa group1 group2 groups
#> 1    7      6      9      B
#> 2    9      8      6      B
#> 3    8      7      4      B
#> 4    3      2      8      A
#> 5    6      5      5      B
#> 6    4      3      2      A
#> 7    5      4      1      B
#> 8    2      1      3      A
scplotter:::gt(group1, group2)
#>   CTaa group1 group2 groups
#> 1    9      8      6      B
#> 2    8      7      4      B
#> 3   10      9      0      B
#> 4    4      3      2      A
#> 5    5      4      1      B
scplotter:::lt(group1, group2)
#>   CTaa group1 group2 groups
#> 1    7      6      9      B
#> 2    3      2      8      A
#> 3    1      0      7      A
#> 4    2      1      3      A
scplotter:::le(group1, group2)
#>   CTaa group1 group2 groups
#> 1    7      6      9      B
#> 2    3      2      8      A
#> 3    6      5      5      B
#> 4    1      0      7      A
#> 5    2      1      3      A
scplotter:::lt(group1, group2, include_zeros = FALSE)
#>   CTaa group1 group2 groups
#> 1    7      6      9      B
#> 2    3      2      8      A
#> 3    2      1      3      A
scplotter:::eq(group1, group2)
#>   CTaa group1 group2 groups
#> 1    6      5      5      B
scplotter:::ne(group1, group2)
#>   CTaa group1 group2 groups
#> 1    7      6      9      B
#> 2    9      8      6      B
#> 3    8      7      4      B
#> 4    3      2      8      A
#> 5   10      9      0      B
#> 6    1      0      7      A
#> 7    4      3      2      A
#> 8    5      4      1      B
#> 9    2      1      3      A

# Use them in a dplyr pipeline
dplyr::mutate(data, Top3 = scplotter:::top(3))
#>    CTaa group1 group2 groups Top3
#> 7     7      6      9      B    7
#> 9     9      8      6      B    9
#> 8     8      7      4      B    8
#> 3     3      2      8      A   NA
#> 6     6      5      5      B   NA
#> 10   10      9      0      B   NA
#> 1     1      0      7      A   NA
#> 4     4      3      2      A   NA
#> 5     5      4      1      B   NA
#> 2     2      1      3      A   NA
dplyr::mutate(data, Top3 = scplotter:::top(3, groups = "groups"))
#>    CTaa group1 group2 groups Top3
#> 7     7      6      9      B    7
#> 9     9      8      6      B    9
#> 8     8      7      4      B    8
#> 3     3      2      8      A    3
#> 6     6      5      5      B   NA
#> 10   10      9      0      B   NA
#> 1     1      0      7      A    1
#> 4     4      3      2      A    4
#> 5     5      4      1      B   NA
#> 2     2      1      3      A   NA
dplyr::mutate(data, Unique = scplotter:::sel(group1 == 0 | group2 == 0))
#>    CTaa group1 group2 groups Unique
#> 7     7      6      9      B     NA
#> 9     9      8      6      B     NA
#> 8     8      7      4      B     NA
#> 3     3      2      8      A     NA
#> 6     6      5      5      B     NA
#> 10   10      9      0      B     10
#> 1     1      0      7      A      1
#> 4     4      3      2      A     NA
#> 5     5      4      1      B     NA
#> 2     2      1      3      A     NA
dplyr::mutate(data, UniqueInG1 = scplotter:::uniq(group1, group2))
#>    CTaa group1 group2 groups UniqueInG1
#> 7     7      6      9      B         NA
#> 9     9      8      6      B         NA
#> 8     8      7      4      B         NA
#> 3     3      2      8      A         NA
#> 6     6      5      5      B         NA
#> 10   10      9      0      B         10
#> 1     1      0      7      A         NA
#> 4     4      3      2      A         NA
#> 5     5      4      1      B         NA
#> 2     2      1      3      A         NA
dplyr::mutate(data, Shared = scplotter:::shared(group1, group2))
#>    CTaa group1 group2 groups Shared
#> 7     7      6      9      B      7
#> 9     9      8      6      B      9
#> 8     8      7      4      B      8
#> 3     3      2      8      A      3
#> 6     6      5      5      B      6
#> 10   10      9      0      B     NA
#> 1     1      0      7      A     NA
#> 4     4      3      2      A      4
#> 5     5      4      1      B      5
#> 2     2      1      3      A      2
dplyr::mutate(data, Greater = scplotter:::gt(group1, group2))
#>    CTaa group1 group2 groups Greater
#> 7     7      6      9      B      NA
#> 9     9      8      6      B       9
#> 8     8      7      4      B       8
#> 3     3      2      8      A      NA
#> 6     6      5      5      B      NA
#> 10   10      9      0      B      10
#> 1     1      0      7      A      NA
#> 4     4      3      2      A       4
#> 5     5      4      1      B       5
#> 2     2      1      3      A      NA
dplyr::mutate(data, Less = scplotter:::lt(group1, group2))
#>    CTaa group1 group2 groups Less
#> 7     7      6      9      B    7
#> 9     9      8      6      B   NA
#> 8     8      7      4      B   NA
#> 3     3      2      8      A    3
#> 6     6      5      5      B   NA
#> 10   10      9      0      B   NA
#> 1     1      0      7      A    1
#> 4     4      3      2      A   NA
#> 5     5      4      1      B   NA
#> 2     2      1      3      A    2
dplyr::mutate(data, LessEqual = scplotter:::le(group1, group2))
#>    CTaa group1 group2 groups LessEqual
#> 7     7      6      9      B         7
#> 9     9      8      6      B        NA
#> 8     8      7      4      B        NA
#> 3     3      2      8      A         3
#> 6     6      5      5      B         6
#> 10   10      9      0      B        NA
#> 1     1      0      7      A         1
#> 4     4      3      2      A        NA
#> 5     5      4      1      B        NA
#> 2     2      1      3      A         2
dplyr::mutate(data, GreaterEqual = scplotter:::ge(group1, group2))
#>    CTaa group1 group2 groups GreaterEqual
#> 7     7      6      9      B           NA
#> 9     9      8      6      B            9
#> 8     8      7      4      B            8
#> 3     3      2      8      A           NA
#> 6     6      5      5      B            6
#> 10   10      9      0      B           10
#> 1     1      0      7      A           NA
#> 4     4      3      2      A            4
#> 5     5      4      1      B            5
#> 2     2      1      3      A           NA
dplyr::mutate(data, Equal = scplotter:::eq(group1, group2))
#>    CTaa group1 group2 groups Equal
#> 7     7      6      9      B    NA
#> 9     9      8      6      B    NA
#> 8     8      7      4      B    NA
#> 3     3      2      8      A    NA
#> 6     6      5      5      B     6
#> 10   10      9      0      B    NA
#> 1     1      0      7      A    NA
#> 4     4      3      2      A    NA
#> 5     5      4      1      B    NA
#> 2     2      1      3      A    NA
dplyr::mutate(data, NotEqual = scplotter:::ne(group1, group2))
#>    CTaa group1 group2 groups NotEqual
#> 7     7      6      9      B        7
#> 9     9      8      6      B        9
#> 8     8      7      4      B        8
#> 3     3      2      8      A        3
#> 6     6      5      5      B       NA
#> 10   10      9      0      B       10
#> 1     1      0      7      A        1
#> 4     4      3      2      A        4
#> 5     5      4      1      B        5
#> 2     2      1      3      A        2

# Compond expressions
dplyr::mutate(data,
  Top3OrEqual = scplotter:::or(scplotter:::top(3), scplotter:::eq(group1, group2)))
#>    CTaa group1 group2 groups Top3OrEqual
#> 7     7      6      9      B           7
#> 9     9      8      6      B           9
#> 8     8      7      4      B           8
#> 3     3      2      8      A          NA
#> 6     6      5      5      B           6
#> 10   10      9      0      B          NA
#> 1     1      0      7      A          NA
#> 4     4      3      2      A          NA
#> 5     5      4      1      B          NA
#> 2     2      1      3      A          NA
dplyr::mutate(data,
  SharedAndGreater = scplotter:::and(
     scplotter:::shared(group1, group2), scplotter:::gt(group1, group2)))
#>    CTaa group1 group2 groups SharedAndGreater
#> 7     7      6      9      B               NA
#> 9     9      8      6      B                9
#> 8     8      7      4      B                8
#> 3     3      2      8      A               NA
#> 6     6      5      5      B               NA
#> 10   10      9      0      B               NA
#> 1     1      0      7      A               NA
#> 4     4      3      2      A                4
#> 5     5      4      1      B                5
#> 2     2      1      3      A               NA